% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_map.R
\name{hex_map}
\alias{hex_map}
\title{Load base gridded hexagonal map for Burkina Faso}
\usage{
hex_map()
}
\value{
a sf object with gridded hexagonal map for Burkina Faso
}
\description{
Function to add the gridded hexagonal map of Burkina Faso.
This function has no argument
}
\examples{
# Plot the map
\dontrun{
library(sf)
library(tidyverse)

# get hex map
hex_bf <-hex_map() # No argument, need internet connection.
ggplot(hex_bf)+geom_sf() # Quick visualization
}
}
