% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbur-package.R
\docType{package}
\name{hexbur-package}
\alias{hexbur}
\alias{hexbur-package}
\title{hexbur: hexbur provides several layers of information for Burkina Faso, particularly useful for analyzing agricultural systems}
\description{
hexbur stands for "hEXagons for BURkina Faso". The basis of this package is a gridded map of Burkina Faso in hexagons of about 340 km2. For each hexagon, different information can be added using the package function. The data is not included in the package itself, but is called up via URL. It therefore requires an Internet connection to operate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oousmane/hexbur}
  \item Report bugs at \url{https://github.com/oousmane/hexbur/issues}
}

}
\author{
\strong{Maintainer}: Ousmane Ouedraogo \email{oousmane@proton.me} (\href{https://orcid.org/0000-0002-6349-4146}{ORCID})

}
\keyword{internal}
