% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_to_usda}
\alias{num_to_usda}
\title{Translation of USDA texture class number to "abr", "name_en" or "name_fr"}
\usage{
num_to_usda(texture = 1, to = "abr")
}
\arguments{
\item{to}{convert numeric USDA class to. one of "abr", "name_en","name_fr"}
}
\value{
translated texture
}
\description{
Translation of USDA texture class number to "abr", "name_en" or "name_fr"
}
\examples{
num_to_usda(texture = 1:12, to = "abr")
num_to_usda(texture = 1:12, to = "name_en")
num_to_usda(texture = 1:12, to = "name_fr")

}
